% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBC_dichotom.R
\name{BBC_cutoff}
\alias{BBC_cutoff}
\title{Bootstrap Cutoff}
\usage{
BBC_cutoff(object)
}
\arguments{
\item{object}{returned value from function \link{BBC_dichotom}}
}
\value{
Function \link{BBC_cutoff} returns a \link[base]{matrix} of bootstrap cutoffs.
}
\description{
..
}
\details{
we use the output of \link{BBC_dichotom}.

but actually this works on the output of \link{optimism_dichotom}.
}
\keyword{internal}
