% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterQp.R
\name{clusterQp}
\alias{clusterQp}
\title{Cluster-Specific Sample Quantiles}
\usage{
clusterQp(
  formula,
  data,
  exclude = character(),
  aggregateQp = mean,
  from = 0.01,
  to = 0.99,
  by = 0.01,
  type = 7,
  ...
)
}
\arguments{
\item{formula}{\link[stats]{formula}, see details in \link[stats]{aggregate.formula}.
End user may use following formulas
to specify the response \eqn{y}, cluster(s) \eqn{c}'s and covariate(s) \eqn{x}'s
\describe{
\item{\code{y ~ 1 | c1}}{cluster \eqn{c_1}, without cluster-specific covariate}
\item{\code{y ~ 1 | c1/c2}}{cluster \eqn{c_1}, and cluster \eqn{c_2} nested in \eqn{c_1}, without cluster-specific covariate}
\item{\code{y ~ x1 + x2 | c1}}{cluster \eqn{c_1}, and cluster-specific covariates \eqn{x_1} and \eqn{x_2}}
\item{\code{y ~ x1 + x2 | c1/c2}}{cluster \eqn{c_1}, cluster \eqn{c_2} nested in \eqn{c_1}, and cluster (\eqn{c_2} at least) specific covariates \eqn{x_1} and \eqn{x_2}}
\item{\code{y ~ .  | c1}}{cluster \eqn{c_1}, and all (supposedly cluster-specific) covariates from \code{data}}
\item{\code{y ~ .  | c1/c2}}{cluster \eqn{c_1}, cluster \eqn{c_2} nested in \eqn{c_1}, and all (supposedly cluster-specific) covariates from \code{data}}
}}

\item{data}{\link[base]{data.frame},
containing
the response \eqn{y}, cluster(s) \eqn{c}'s and covariate(s) \eqn{x}'s}

\item{exclude}{(optional) \link[base]{character} \link[base]{vector},
variable(s) excluded from aggregation,
e.g., use \code{exclude = c('z1', 'z2')}
to exclude variables \eqn{z_1} and \eqn{z_2}}

\item{aggregateQp}{\link[base]{function} used to aggregate the \link[stats]{quantile}s from
cluster \eqn{c_2} (if present). Choices include \link[base]{mean}, \link[base]{max} and \link[base]{min}.}

\item{from, to, by}{\link[base]{double} scalars,
the starting, end, and increment values
to specify a \link[base]{seq}uence of probabilities
\eqn{p = (p_1,\cdots,p_N)'}
for the sample \link[stats]{quantile}s \eqn{q = (q_1,\cdots,q_N)'}}

\item{type}{\link[base]{integer} scalar, \code{type} of \link[stats]{quantile} algorithm}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{clusterQp} returns an \link[stats]{aggregate}d \link[base]{data.frame}.
A \link[base]{double} \link[base]{matrix} of \eqn{N} columns is created to store
the sample \link[stats]{quantile}s \eqn{q} of each \link[stats]{aggregate}d cluster.
The column names of this \link[stats]{quantile} \link[base]{matrix} are the probabilities \eqn{p}.
}
\description{
Sample \link[stats]{quantile}s in each cluster of observations.
}
\details{
Function \link{clusterQp} calculates \eqn{N} sample \link[stats]{quantile}s
in each \link[stats]{aggregate}d cluster of observations.
}
\examples{
# see ?`Qindex-package` for examples
}
\keyword{internal}
