% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBC_dichotom.R
\name{coxph_optQ}
\alias{coxph_optQ}
\title{..}
\usage{
coxph_optQ(
  formula,
  data,
  optQ,
  dQ = dichotom_optQ(edp = eval(formula[[2L]], envir = data), optQ = data[[optQ]])
)
}
\arguments{
\item{formula}{..}

\item{data}{..}

\item{optQ}{\link[base]{character} scalar}

\item{dQ}{returned object from \link{dichotom_optQ}}
}
\value{
\link{coxph_optQ} returns a \link[base]{numeric} vector of the concatenated
coefficients of \strong{univariable model} (which is returned in attribute \code{'ucox'})
and coefficients \strong{multivariable model} (which is returned in attribute \code{'mcox'}).
}
\description{
..
}
