% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{FRindex}
\alias{predict.FRindex}
\alias{optim_splitSample_dichotom}
\title{Back Compatibility}
\usage{
FRindex(...)

\method{predict}{FRindex}(...)

optim_splitSample_dichotom(...)
}
\arguments{
\item{...}{parameters that have been \link[base]{.Defunct}.}
}
\description{
Functions that have been \link[base]{.Defunct}.
}
\keyword{internal}
