% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBC_dichotom.R
\name{dichotom_int}
\alias{dichotom_int}
\title{Dichotomizing}
\usage{
dichotom_int(edp, contX)
}
\arguments{
\item{edp}{\link[survival]{Surv}, \link[base]{logical}, or \link[base]{numeric} object, the endpoint}

\item{contX}{see \link{BBC_dichotom}}
}
\value{
\link{dichotom_int} returns a \link[base]{logical} \link[base]{matrix} of
the same dimension and dimension names as argument \code{contX},
as dichotomized using the first node of \link[rpart]{rpart} as threshold.
The thresholds (per column) are returned as an \link[base]{numeric} vector
in attribute \code{'thres'}.
}
\description{
Generate dichotimized variables by a threshold calculated from \link{rpart}
using \link[survival]{Surv}, \link[base]{logical}, or \link[base]{numeric} endpoint and each quantile
}
