% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBC_dichotom.R
\name{dichotom_optQ}
\alias{dichotom_optQ}
\title{Helper function \link{dichotom_optQ}}
\usage{
dichotom_optQ(edp, optQ, control = rpart.control(minbucket = 40))
}
\arguments{
\item{edp}{\link[survival]{Surv} object, the endpoint}

\item{optQ}{\link[base]{matrix} of optimal quantiles}

\item{control}{see \link[rpart]{rpart}}
}
\value{
\link{dichotom_optQ} returns a \link[base]{logical} \link[base]{matrix} of
the same dimension and dimension names as argument \code{optQ},
as dichotomized using the first node of \link[rpart]{rpart} as threshold.
The thresholds (per column) are returned as an \link[base]{numeric} vector
in attribute \code{'thres'}.
}
\description{
..
}
