% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRidx.R
\docType{class}
\name{FRidx}
\alias{FRidx}
\alias{FRidx-class}
\alias{FRidx_prefit_}
\title{Functional Regression Indices}
\usage{
FRidx(formula, data, sign_prob = 0.5, ...)

FRidx_prefit_(
  formula,
  data,
  family,
  nonlinear = FALSE,
  knot_pct = 0.4,
  knot_value = ceiling(nxgrid * knot_pct),
  ...
)
}
\arguments{
\item{formula}{\link[stats]{formula}, e.g., \code{y~X}.
Types of response \eqn{y} may be \link[base]{double}, \link[base]{logical} and \link[survival]{Surv}.
Functional predictor \eqn{X} is a tabulated \link[base]{double} \link[base]{matrix};
the rows of \eqn{X} correspond to the subjects,
while the columns of \eqn{X} correspond to a \emph{common tabulating grid} shared by all subjects.
The \link[base]{numeric} values of the grid are in the \link[base]{colnames} of \eqn{X}}

\item{data}{\link[base]{data.frame}, e.g., returned object from function \link{clusterQp}}

\item{sign_prob}{\link[base]{double} scalar between 0 and 1,
probability corresponding to
the selected nearest-even \link[stats]{quantile} in the grid,
which is used to determine the \link[base]{sign} of the (non)linear functional regression indices.
Default is \code{.5}, i.e., the nearest-even \link[stats]{median} of the grid}

\item{...}{additional parameters, currently not in use.}

\item{family}{\link[stats]{family} object,
see function \link[mgcv]{gam}.
Default values are
\itemize{
\item \code{mgcv::cox.ph()} for \link[survival]{Surv} response \eqn{y};
\item \code{binomial(link = 'logit')} for \link[base]{logical} response \eqn{y};
\item \code{gaussian(link = 'identity')} for \link[base]{double} response \eqn{y}
}}

\item{nonlinear}{\link[base]{logical} scalar,
whether to use nonlinear or linear functional regression model.
Default \code{FALSE}}

\item{knot_pct}{(only when \code{nonlinear = FALSE})
positive \link[base]{double} scalar,
percentage of the number of columns of \eqn{X},
to be used as \code{knot_value}.
Default is \eqn{40\%}.
If \code{knot_value} is provided by the end-user, then \code{knot_pct} is ignored.}

\item{knot_value}{(only when \code{nonlinear = FALSE})
positive \link[base]{integer} scalar, number of knots
(i.e., parameter \code{k} in the spline smooth function \link[mgcv]{s})
used in \link[mgcv]{gam}.
Default is the \link[base]{ceiling} of \code{knot_pct} of
the column dimension of \eqn{X}}
}
\value{
Function \link{FRidx} returns an \linkS4class{FRidx} object,
which is an instance of an \link[base]{S4} class.
See section \strong{Slots} for details.
}
\description{
Functional regression indices
based on linear or nonlinear functional predictors.
}
\details{
Function \link{FRidx} calculates
the functional regression indices in the following steps.

\enumerate{

\item Fit a functional regression model (via \link[mgcv]{gam})
of response \eqn{y}
with functional predictor \eqn{X};

\item Obtain the \link[base]{sign} of the \link[stats]{cor}relation between
\itemize{
\item {\eqn{X_{\cdot,j}},
the user-selected \eqn{j}-th column of functional predictor \eqn{X}.
By default, this is the column corresponding to
the \link[stats]{median} of the grid;}
\item \code{gam(.)$linear.predictors}, from Step 1
}

}

\emph{Functional regression indices} (slot \verb{@.Data} of returned object)
are the product of
\code{sign} (from Step 2) and \code{gam(.)$linear.predictors} (from Step 1).
Multiplication by \code{sign} ensures
that the functional regression indices
are positively correlated with the user-selected \eqn{X_{\cdot,j}}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{double} \link[base]{vector},
functional regression indices, see section \strong{Details}.}

\item{\code{formula}}{see section \strong{Arguments}}

\item{\code{gam}}{a \link[mgcv]{gam} object}

\item{\code{gpf}}{a \code{'gam.prefit'} object, which is the returned object
from function \link[mgcv]{gam} with argument \code{fit = FALSE}}

\item{\code{p.value}}{\link[base]{numeric} scalar,
\eqn{p}-value for the test of significance of the functional predictor,
based on slot \verb{@gam}}

\item{\code{sign}}{\link[base]{double} scalar of either 1 or -1,
see Step 2 in section \strong{Details}.}
}}

\examples{
# see ?`Qindex-package`
}
\references{
Cui, E., Crainiceanu, C. M., & Leroux, A. (2021).
Additive Functional Cox Model. Journal of Computational and Graphical Statistics.
\doi{10.1080/10618600.2020.1853550}

Gellar, J. E., Colantuoni, E., Needham, D. M., & Crainiceanu, C. M. (2015).
Cox regression models with functional covariates for survival data. Statistical Modelling.
\doi{10.1177/1471082X14565526}
}
