% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRindex.R
\docType{class}
\name{FRindex-class}
\alias{FRindex-class}
\title{\linkS4class{FRindex} Object: Functional Regression Index Model}
\description{
An S4 class for the functional regression index model.
}
\section{Slots}{

\describe{
\item{\code{gam}}{\link[mgcv]{gam} object}

\item{\code{sign}}{\link[base]{numeric} scalar}

\item{\code{FRi}}{\link[base]{numeric} \link[base]{vector}, functional regression index}

\item{\code{FRwt}}{\link[base]{numeric} \link[base]{vector}, functional regression weights}

\item{\code{xarg}}{\link[base]{numeric} \link[base]{vector}, \eqn{x}-domain}

\item{\code{X}}{\link[base]{numeric} \link[base]{matrix},
each row representing the tabulated, on a common grid,
functional predictor values for each subject.}
}}

