% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrandSurface.R
\name{integrandSurface}
\alias{integrandSurface}
\title{Integrand Surface(s) for One or More \linkS4class{FRidx} Model(s)}
\usage{
integrandSurface(
  ...,
  newdata = data,
  n = 501L,
  subj_vis = seq_len(min(50L, .row_names_info(newdata, type = 2L))),
  ylim = range(X, newX)
)
}
\arguments{
\item{...}{one or more \linkS4class{FRidx} objects based on the training set.}

\item{newdata}{test \link[base]{data.frame}, with at least
the response \eqn{y^{\text{new}}} and
the \link[base]{double} \link[base]{matrix} of
functional predictor values \eqn{X^{\text{new}}}
of the test set, tabulated on the same \eqn{x}-grid as the training set.
If missing, the training set will be used.}

\item{n}{\link[base]{integer} scalar, fineness of visualization,
default \code{501L}. See parameter \code{n.grid} of function \link[mgcv]{vis.gam}.}

\item{subj_vis}{\link[base]{integer} scalar or \link[base]{vector},
row indices of \code{newdata} to be visualized.
Default \code{1:2}, i.e., the first two test subjects.
Use \code{subj_vis = NULL} to disable visualization of \code{newdata}.}

\item{ylim}{..}
}
\value{
Function \link{integrandSurface} returns a pretty \link[plotly]{plotly} object
to showcase the \link[graphics]{persp}ective plot of the
\emph{\strong{integrand surface}} of functional regression indices.
The \emph{\strong{integrand curves}} of selected test subjects will also be displayed.
}
\description{
..
}
\details{
Function \link{integrandSurface} ..
}
