% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBC_dichotom.R
\name{model_dichotom}
\alias{model_dichotom}
\title{Model with Dichotomized Predictors}
\usage{
model_dichotom(
  formula,
  data,
  contX,
  dQ = dichotom_int(edp = eval(formula[[2L]], envir = data), contX = data[[contX]])
)
}
\arguments{
\item{formula}{same formula in \link{BBC_dichotom}}

\item{data}{\link[base]{data.frame}}

\item{contX}{see \link{BBC_dichotom}}

\item{dQ}{returned object from \link{dichotom_int}}
}
\value{
\link{model_dichotom} returns a \link[base]{numeric} vector of the concatenated
coefficients.
}
\description{
fit Cox proportional hazard model, logistic regression, or linear regression with
dichotomized marker and/or other covariates built as formula in \link{BBC_dichotom}
}
