% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimSplit_dichotom.R
\name{optimSplit_dichotom}
\alias{optimSplit_dichotom}
\alias{split_dichotom}
\alias{splits_dichotom}
\alias{quantile.splits_dichotom}
\title{Optimal Dichotomizing Predictors via Repeated Sample Splits}
\usage{
optimSplit_dichotom(
  formula,
  data,
  include = quote(p1 > 0.15 & p1 < 0.85),
  top = 1L,
  nsplit,
  ...
)

split_dichotom(y, x, train, ...)

splits_dichotom(y, x, trains = rSplit(y, ...), ...)

\method{quantile}{splits_dichotom}(x, probs = 0.5, ...)
}
\arguments{
\item{formula, y, x}{two-sided \link[stats]{formula} \code{y~X} or \code{y~x1+x2}.
Types of response \eqn{y} may be \link[base]{double}, \link[base]{logical} and \link[survival]{Surv}.
Candidate \link[base]{numeric} predictors \eqn{x}'s may be specified as the columns of one \link[base]{matrix} column, e.g., \code{y~X}; or as several \link[base]{vector} columns, e.g., \code{y~x1+x2}.
In helper functions, \code{x} is a \link[base]{numeric} \link[base]{vector}.}

\item{data}{\link[base]{data.frame}}

\item{include}{(optional) \link[base]{language}, inclusion criteria for the optimal dichotomizing predictors.
Default \code{(p1>.15 & p1<.85)} specifies a user-desired range of \eqn{p_1}.
See explanation of \eqn{p_1} in section \strong{Returns of Helper Functions}.}

\item{top}{positive \link[base]{integer} scalar, number of optimal dichotomizing predictors, default \code{1L}}

\item{nsplit, ...}{additional parameters for function \link{rSplit}}

\item{train}{\link[base]{logical} \link[base]{vector} for helper function \link{split_dichotom}, indices of training (\code{TRUE}) and test (\code{FALSE}) subjects}

\item{trains}{(optional) \link[base]{list} of \link[base]{logical} \link[base]{vector}s for helper function \link{splits_dichotom}, indices of multiple training-test sample splits.}

\item{probs}{\link[base]{double} scalar for helper function \link{quantile.splits_dichotom}, see \link[stats]{quantile}}
}
\value{
Function \link{optimSplit_dichotom} returns an object of \link[base]{class} \code{'optimSplit_dichotom'},
which is a \link[base]{list} of dichotomizing \link[base]{function}s
of \link[base]{length}-\code{top}.
The returned value has \link[base]{attributes},
\describe{
\item{\code{attr(,'formula')}}{
input \link[stats]{formula}}
\item{\code{attr(,'data')}}{
input \link[base]{data.frame}}
}
}
\description{
Functions explained in this documentation are,

\describe{
\item{\link{optimSplit_dichotom}}{to identify the optimal dichotomizing predictors using repeated sample splits.}
\item{\link{split_dichotom}}{helper function to perform a univariable regression model on the test set with a dichotomized predictor, using a dichotomizing rule determined by a recursive partitioning of the training set.}
\item{\link{quantile.splits_dichotom}}{helper function to locate a quantile of multiple \link{split_dichotom} objects, based on the estimated univariable regression coefficient.}
}
}
\details{
Function \link{optimSplit_dichotom} selects the optimal dichotomizing predictors via repeated sample splits
in the following steps,

\enumerate{

\item Generate multiple training-test sample splits, see \link{rSplit}

\item For each candidate predictor \eqn{x_i},
find the median \link{split_dichotom}
(using helper function \link{quantile.splits_dichotom})
of the multiple sample splits from Step 1.

\item (Optional) limit the selection in a subset of the candidate predictors to guarantee a user-desired range of \eqn{p_1} (see explanations in section \strong{Returns of Helper Functions});

\item Rank the candidate predictors, from either Step 2 or Step 3,
by the decreasing order of the \link[base]{abs}olute values of
the univariable regression coefficient estimate of the corresponding \link{split_dichotom} objects.

}

The \emph{optimal dichotomizing predictors} are the ones
with the largest \link[base]{abs}olute values of
the estimated univariable regression coefficients
of the corresponding \link{split_dichotom} objects.
}
\section{Details on Helper Functions}{

\subsection{Univariable regression model with a split-dichotomized predictor}{

Helper function \link{split_dichotom} performs a univariable regression model on the test set with a dichotomized predictor, using a dichotomizing rule determined by a recursive partitioning of the training set.
Specifically, given a training-test sample split,
\enumerate{
\item find the \emph{dichotomizing rule} \eqn{\mathcal{D}} of the predictor \eqn{x_0} given the response \eqn{y_0} in the training set, see \link{rpartD};
\item fit a univariable regression model of the response \eqn{y_1} with the dichotomized predictor \eqn{\mathcal{D}(x_1)} in the test set.
}
Currently the Cox proportional hazards (\link[survival]{coxph}) regression for \link[survival]{Surv} response, logistic (\link[stats]{glm}) regression for \link[base]{logical} response and linear (\link[stats]{lm}) regression for \link[stats]{gaussian} response are supported.
}


\subsection{List of \link{split_dichotom}s, based on multiple training-test sample splits}{

Helper function \link{splits_dichotom} executes function \link{split_dichotom} on the response \eqn{y} and predictor \eqn{x}, based on multiple training-test sample splits.
}


\subsection{Quantile of \link{splits_dichotom}}{

Helper function \link{quantile.splits_dichotom} is a method dispatch of the S3 generic function \link[stats]{quantile} on \link{splits_dichotom} object.
Specifically,

\enumerate{
\item {collect the univariable regression coefficient estimate from each \link{split_dichotom} element in the input \link{splits_dichotom} object;}
\item {find the nearest-even (i.e., \code{type = 3}) \link[stats]{quantile} of the coefficients from Step 1. By default, we use the \link[stats]{median} (i.e., \code{prob = .5});}
\item {the \link{split_dichotom} element corresponding to the selected coefficient quantile in Step 2, is returned.}
}
}
}

\section{Returns of Helper Functions}{


Helper functions \link{split_dichotom} returns a \link{split_dichotom} object, which is either a Cox proportional hazards (\link[survival]{coxph}), a logistic (\link[stats]{glm}), or a linear (\link[stats]{lm}) regression model, with additional \link[base]{attributes}

\describe{
\item{\code{attr(,'rule')}}{\link[base]{function}, dichotomizing rule \eqn{\mathcal{D}} based on the training set}
\item{\code{attr(,'text')}}{\link[base]{character} scalar, human-friendly description of \eqn{\mathcal{D}}}
\item{\code{attr(,'p1')}}{\link[base]{double} scalar, \eqn{p_1 = \text{Pr}(\mathcal{D}(x_1)=1)}}
\item{\code{attr(,'coef')}}{\link[base]{double} scalar, univariable regression coefficient estimate of \eqn{y_1\sim\mathcal{D}(x_1)}}
}



Helper functions \link{splits_dichotom} returns a \link[base]{list} of \link{split_dichotom} objects.


Helper function \link{quantile.splits_dichotom} returns a \link{split_dichotom} object.
}

