% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optQp.R
\name{optQp}
\alias{optQp}
\title{Optimal Quantile Predictor}
\usage{
optQp(x, n = 3L, ...)
}
\arguments{
\item{x}{\link{eval_dichotom} object}

\item{n}{\link[base]{integer} scalar, number of optimal quantiles to be printed}

\item{...}{additional parameters, not currently in use}
}
\value{
\link{optQp} returns a \link[base]{data.frame} with designated number of optimal quantiles
}
\description{
Wrapper of \link{eval_dichotom}.

To create a \link[base]{data.frame} with designated number of optimal quantiles returned from \link{eval_dichotom}
in addition to \code{subjID}, \link[survival]{Surv} endpoint and
\link[base]{numeric}, \link[base]{character}, or \link[base]{logical} predictors
}
\examples{
# see ?BBC_dichotom

}
