% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.FRidx.R
\name{predict.FRidx}
\alias{predict.FRidx}
\title{Predicted Functional Regression Indices}
\usage{
\method{predict}{FRidx}(object, newdata = object@gam$data, ...)
}
\arguments{
\item{object}{an \linkS4class{FRidx} object based on the training set.}

\item{newdata}{test \link[base]{data.frame}, with at least
the response \eqn{y^{\text{new}}} and
the \link[base]{double} \link[base]{matrix} of
functional predictor values \eqn{X^{\text{new}}}
of the test set, tabulated on the same \code{object@xgrid} as the training set.
If missing, the training set \code{object@gam$data} will be used.}

\item{...}{additional parameters, currently not in use.}
}
\value{
Function \link{predict.FRidx} returns a
\link[base]{double} \link[base]{vector},
which is the predicted functional regression indices on the test set.
}
\description{
To predict functional regression indices of a test set.
}
\details{
Function \link{predict.FRidx} computes
the predicted functional regression indices on the test set,
which is
the product of function \link[mgcv]{predict.gam} return
and the correlation sign based on training set
(\code{object@sign}, see Step 3 of section \strong{Details} of function \link{FRidx}).
Multiplication by \code{object@sign} is required to ensure
that the predicted functional regression indices
are positively associated with the \strong{training} functional predictor values
at the selected quantile of \code{object@xgrid}.
}
