% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimSplit_dichotom.R
\name{predict.optimSplit_dichotom}
\alias{predict.optimSplit_dichotom}
\title{predict.optimSplit_dichotom}
\usage{
\method{predict}{optimSplit_dichotom}(
  object,
  formula = attr(object, which = "formula", exact = TRUE),
  newdata = attr(object, which = "data", exact = TRUE),
  ...
)
}
\arguments{
\item{object}{an \link{optimSplit_dichotom} object}

\item{formula}{\link[stats]{formula}}

\item{newdata}{(optional) test \link[base]{data.frame}.
If missing, the training data is used}

\item{...}{..}
}
\value{
Function \link{predict.optimSplit_dichotom} returns ..
}
\description{
..
}
