% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optQp.R
\name{print.eval_dichotom}
\alias{print.eval_dichotom}
\title{Print \link{eval_dichotom} Object}
\usage{
\method{print}{eval_dichotom}(x, n = 3L, ...)
}
\arguments{
\item{x}{\link{eval_dichotom} object}

\item{n}{\link[base]{integer} scalar, number of optimal quantiles to be printed}

\item{...}{additional parameters, currently not in use}
}
\value{
\link{print.eval_dichotom} does not have a returned value.
}
\description{
To print the optimal quantile that has the largest absolute value of coefficients
for a given \link[survival]{Surv} outcome, \link[base]{logical}, or \link[base]{numeric} outcome
}
\details{
\link{print.eval_dichotom} simply calls \link{summary.eval_dichotom} and
print three largest absolute value of coefficients, hazard ratio or odds ratio, and threshold
}
