% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optQp.R
\name{print.evalQp}
\alias{print.evalQp}
\title{Print \link{evalQp} Object}
\usage{
\method{print}{evalQp}(x, n = 3L, ...)
}
\arguments{
\item{x}{..}

\item{n}{..}

\item{...}{..}
}
\value{
\link{print.evalQp} does not have a returned value.
}
\description{
..
}
\details{
\link{print.evalQp} simply calls \link{summary.evalQp} ..
}
