% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{Qindex-package}
\alias{Qindex}
\alias{Qindex-package}
\title{Continuous and Dichotomized Index Predictors Based on Distribution Quantiles}
\description{
Continuous and dichotomized index predictors based on distribution quantiles.
}
\section{Step 1. Cluster-Specific Sample Quantiles \link{clusterQp}}{


Function \link{clusterQp} calculates user-selected sample quantiles in each cluster of observations.
}

\section{Step 2. Three Types of Index Biomarker (??)}{

\subsection{Optimal Dichotomizing Predictor(s) Selection via Dichotomizing Split Sample \link{optimSplit_dichotom}}{

Function \link{optimSplit_dichotom} identifies
the optimal dichotomizing predictors using repeated sample splits on the \emph{training set}.

Function \link{predict.optimSplit_dichotom} .. testing set
}

\subsection{Functional Regression Indices \link{FRidx}}{

Function \link{FRidx} with option \code{nonlinear = FALSE} (default) .. training set

Function \link{predict.FRidx} .. testing set

Function \link{FRidx} with option \code{nonlinear = TRUE} .. training set

Function \link{predict.FRidx} .. testing set
}
}

\section{Step 3. Bootstrap-Based Optimism Correction \link{BBC_dichotom}}{


Function \link{BBC_dichotom}: Bootstrap-based optimism correction for dichotomizing selected predictor(s).
}

\examples{
### Data Preparation

library(survival)
data(Ki67, package = 'Qindex.data')
Ki67c = within(Ki67[complete.cases(Ki67), , drop = FALSE], expr = {
  marker = log1p(Marker); Marker = NULL
  PFS = Surv(RECFREESURV_MO, RECURRENCE)
})
(npt = length(unique(Ki67c$PATIENT_ID))) # 592

### Step 1: Cluster-Specific Sample Quantiles

Ki67q = clusterQp(marker ~ . - tissueID - inner_x - inner_y | PATIENT_ID, data = Ki67c)
stopifnot(is.matrix(Ki67q$marker))

set.seed(234); id = sort.int(sample.int(n = npt, size = 480L))
Ki67q_0 = Ki67q[id, , drop = FALSE] # training set
Ki67q_1 = Ki67q[-id, , drop = FALSE] # test set

### Step 2 (after Step 1)

## Step 2a: Linear Functional Regression Indices
(fr = FRidx(PFS ~ marker, data = Ki67q_0))
stopifnot(all.equal.numeric(c(fr), predict(fr)))
integrandSurface(fr)
fr_1 = predict(fr, newdata = Ki67q_1)
integrandSurface(fr, newdata = Ki67q_1)

## Step 2b: Non-Linear Functional Regression Indices
(nlfr = FRidx(PFS ~ marker, data = Ki67q_0, nonlinear = TRUE))
stopifnot(all.equal.numeric(c(nlfr), predict(nlfr)))
integrandSurface(nlfr)
if (FALSE) {
  integrandSurface(nlfr, subj_vis = NULL)
  vis.gam(nlfr@gam)
  #autoplot.vis_gam_(vis_gam_(nlfr@gam))
} # these are mathematically identical
nlfr_1 = predict(nlfr, newdata = Ki67q_1)
integrandSurface(nlfr, newdata = Ki67q_1)

### Step 2c: Optimal Dichotomizing
set.seed(14837); (m1 = optimSplit_dichotom(
  PFS ~ marker, data = Ki67q_0, nsplit = 20L, 
  include = (highX > .15 & highX < .85), top = 2L)) 
predict(m1)
predict(m1, newdata = Ki67q_1) # using both quantile selected AND cutoff identified.

### Step 3 (after Step 1 & 2)

Ki67q_0a = within.data.frame(Ki67q_0, expr = {
  FR = std_IQR(fr) 
  nlFR = std_IQR(nlfr)
  optS = std_IQR(marker[,'0.27'])
})
Ki67q_1a = within.data.frame(Ki67q_1, expr = {
  FR = std_IQR(fr_1)
  nlFR = std_IQR(nlfr_1)
  optS = std_IQR(marker[,'0.27']) 
})
# `optS`: use the best quantile but discard the cutoff identified by [optimSplit_dichotom]
# all models below can also be used on training data `Ki67q_0a`
# naive use
summary(coxph(PFS ~ NodeSt + Tstage + FR, data = Ki67q_1a))
summary(coxph(PFS ~ NodeSt + Tstage + nlFR, data = Ki67q_1a))
summary(coxph(PFS ~ NodeSt + Tstage + optS, data = Ki67q_1a))
# set.seed if necessary
summary(BBC_dichotom(PFS ~ NodeSt + Tstage, dichotom = ~ FR, data = Ki67q_1a))
# set.seed if necessary
summary(BBC_dichotom(PFS ~ NodeSt + Tstage, dichotom = ~ nlFR, data = Ki67q_1a))
# set.seed if necessary
summary(BBC_dichotom(PFS ~ NodeSt + Tstage, dichotom = ~ optS, data = Ki67q_1a))
}
\references{
Selection of optimal quantile protein biomarkers based on cell-level immunohistochemistry data.
Misung Yi, Tingting Zhan, Amy P. Peck, Jeffrey A. Hooke, Albert J. Kovatich, Craig D. Shriver,
Hai Hu, Yunguang Sun, Hallgeir Rui and Inna Chervoneva. BMC Bioinformatics, 2023. \doi{10.1186/s12859-023-05408-8}

Quantile index biomarkers based on single-cell expression data.
Misung Yi, Tingting Zhan, Amy P. Peck, Jeffrey A. Hooke, Albert J. Kovatich, Craig D. Shriver,
Hai Hu, Yunguang Sun, Hallgeir Rui and Inna Chervoneva.
Laboratory Investigation, 2023. \doi{10.1016/j.labinv.2023.100158}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Misung Yi \email{misung.yi@dankook.ac.kr} (\href{https://orcid.org/0000-0002-4007-5408}{ORCID}) [copyright holder]
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID}) [copyright holder]
}

}
