% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{Qindex_cv}
\alias{Qindex_cv}
\title{Qindex via Cross Validation}
\usage{
Qindex_cv(formula, data, k, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{data}{\link[base]{data.frame}}

\item{k}{\link[base]{integer} scalar}

\item{...}{additional parameters, currently not in use}
}
\description{
Qindex via Cross Validation
}
\examples{
data(wrobel_lung, package = 'spatstat.grouped.data')
library(spatstat.grouped)
lungQp = grouped_ppp(hladr ~ OS | patient_id/image_id, f_sum_ = 'min', data = wrobel_lung) |> 
 aggregate_quantile(by = ~ patient_id, probs = seq.int(from = .05, to = .95, by = .01))

lung_CV_QI = Qindex_cv(OS ~ hladr.quantile, data = lungQp, k = 5L, nonlinear = FALSE)
head(lung_CV_QI)   
boxplot(QI ~ folds., data=lung_CV_QI) 
library(survival)
summary(coxph(OS ~ QI, data = lung_CV_QI))

}
