% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRidx.R
\name{show,FRidx-method}
\alias{show,FRidx-method}
\title{Show \linkS4class{FRidx} Object}
\usage{
\S4method{show}{FRidx}(object)
}
\arguments{
\item{object}{an \linkS4class{FRidx} object}
}
\value{
The \link[base]{S4} \link[methods]{show} method of \linkS4class{FRidx} object
does not have a returned value.
}
\description{
Show \linkS4class{FRidx} object.
}
\keyword{internal}
