% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRindex.R
\name{show,FRindex-method}
\alias{show,FRindex-method}
\title{Show \linkS4class{FRindex} Object}
\usage{
\S4method{show}{FRindex}(object)
}
\arguments{
\item{object}{an \linkS4class{FRindex} object, returned from function \code{\link[=FRindex]{FRindex()}}}
}
\value{
The \link[base]{S4} \link[methods]{show} method of \linkS4class{FRindex} object returns a \link[base]{data.frame}.
This is the input \code{data} with the functional predictor values \eqn{X} removed,
and the functional regression indices \verb{@index} appended.
}
\description{
Show \linkS4class{FRindex} object.
}
\keyword{internal}
