% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qindex.R
\name{show,Qindex-method}
\alias{show,Qindex-method}
\title{Show \linkS4class{Qindex} Object}
\usage{
\S4method{show}{Qindex}(object)
}
\arguments{
\item{object}{an \linkS4class{Qindex} object}
}
\value{
The \link[base]{S4} \link[methods]{show} method of \linkS4class{Qindex} object
does not have a returned value.
}
\description{
Show \linkS4class{Qindex} object.
}
\keyword{internal}
