% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0std_.R
\name{std_IQR}
\alias{std_IQR}
\title{Standardization via \link[stats]{median} and \link[stats]{IQR}}
\usage{
std_IQR(x)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}
}
\value{
\link{std_IQR} returns a \link[base]{numeric} \link[base]{vector} of the same length as \code{x}
}
\description{
Standardization via \link[stats]{median} and \link[stats]{IQR}
}
\examples{
std_IQR(rnorm(20))

}
