% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0std_.R
\name{std_mad}
\alias{std_mad}
\title{Standardization via \link[stats]{median} and \link[stats]{mad}}
\usage{
std_mad(x)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}
}
\value{
\link{std_mad} returns a \link[base]{numeric} \link[base]{vector} of the same length as \code{x}
}
\description{
Standardization via \link[stats]{median} and \link[stats]{mad}.
}
\examples{
std_mad(rnorm(20))

}
