% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optQp.R
\name{summary.eval_dichotom}
\alias{summary.eval_dichotom}
\title{Summary Information of \link{eval_dichotom} Object}
\usage{
\method{summary}{eval_dichotom}(object, FUN = median, ...)
}
\arguments{
\item{object}{\link{eval_dichotom} object}

\item{FUN}{summarizing \link[base]{function}, either \link[stats]{median} (default) or \link[base]{mean}}

\item{...}{additional parameters, currently not in use}
}
\value{
\link{summary.eval_dichotom} returns a \link[base]{data.frame} with three (3) columns
\describe{
\item{\code{abs.coef}}{effect sizes (absolute value of log hazard ratio, log odds ratio, or coefficient)}
\item{\code{HR}}{corresponding hazard ratio or odds ratio}
\item{\code{threshold}}{threshold for dichotomizing the quantile}
}
}
\description{
To summarize \link{eval_dichotom} object
}
\details{
\link{summary.eval_dichotom} present effect sizes (absolute value of log hazard ratio, log odds ratio, or coefficient), corresponding hazard ratio or odds ratio, and threshold to dichotomize quantile.
And then sort by \code{abs.coef}
}
\seealso{
\link[base]{summary}
}
