% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optQp.R
\name{summary.evalQp}
\alias{summary.evalQp}
\title{Summary Information of \link{evalQp} Object}
\usage{
\method{summary}{evalQp}(object, FUN = median, ...)
}
\arguments{
\item{object}{\link{evalQp} object}

\item{FUN}{\link[stats]{median} (default) or \link[base]{mean}}

\item{...}{..}
}
\value{
\link{summary.evalQp} summarize ...
}
\description{
Summary method for \link{optQp} Object
}
