% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0bootIDX.R
\name{bootIDX}
\alias{bootIDX}
\title{Bootstrap Indexes}
\usage{
bootIDX(N, ...)
}
\arguments{
\item{N}{positive \link[base]{integer} scalar, sample size}

\item{...}{additional parameters of \link[boot]{boot}, most importantly the number of bootstrap replicates \code{R}}
}
\value{
\link{bootIDX} returns a \link[base]{list} of \link[base]{integer} \link[base]{vector}s.
Each element is the indexes of a boot strap sample.
}
\description{
Generate a \emph{series of} bootstrap indexes.
}
\examples{
bootIDX(10L, R = 5L)

}
