% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_doc.R
\docType{data}
\name{celldata}
\alias{celldata}
\alias{Ki67}
\alias{PR}
\title{Ki67 and PR Data}
\format{
\describe{
\item{\code{PATIENT_ID}}{\link[base]{factor}, ..}
\item{\code{RECURRENCE}}{\link[survival]{Surv}, ...}
\item{\code{Marker}}{\link[base]{numeric}, ...}
\item{\code{Array}}{\link[base]{character}, ..}
\item{all other columns}{...}
}

An object of class \code{data.frame} with 467603 rows and 7 columns.
}
\usage{
Ki67

PR
}
\description{
Ki67 cell data containing 623 patients and PR cell data containing 631 patients.
A total of 572 patients are in common.
}
\keyword{datasets}
