% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRidx.R
\docType{class}
\name{FRidx}
\alias{FRidx}
\alias{FRidx-class}
\alias{FRidx_prefit_}
\title{Functional Regression Indices}
\usage{
FRidx(formula, data, sign_prob = 0.5, ...)

FRidx_prefit_(
  formula,
  data,
  family,
  nonlinear = FALSE,
  knot_pct = 0.4,
  knot_value = ceiling(nxgrid * knot_pct),
  ...
)
}
\arguments{
\item{formula}{a two-sided \link[stats]{formula}.
\describe{
\item{Left-hand-side}{is the \link[base]{name} of the response \eqn{y}.
Supported types of responses are \link[base]{double}, \link[base]{logical} and \link[survival]{Surv}.}
\item{Right-hand-side}{is the \link[base]{name}
of the tabulated \link[base]{double} \link[base]{matrix} \eqn{X} of functional predictor values.
Each row of \eqn{X} represents the tabulated values for a subject.
All rows/subjects are tabulated on a common \code{xgrid}.
Each column of \eqn{X} represents the tabulated values at a point on the common \code{xgrid} for each subject.}
}}

\item{data}{\link[base]{data.frame} (e.g., returned from function \link{clusterQp}), with
the response \eqn{y} and the tabulated functional predictor values \eqn{X}
specified in \code{formula}.}

\item{sign_prob}{\link[base]{double} scalar between 0 and 1,
probability corresponding to the selected nearest-even quantile in \code{xgrid},
which is used to determine the \link[base]{sign} of the (nonlinear) functional regression indices.
Default is \code{.5}, i.e., the nearest-even \link[stats]{median} of \code{xgrid}}

\item{...}{additional parameters, currently not in use.}

\item{family}{\link[stats]{family} object,
see function \link[mgcv]{gam}.
Default values are
\itemize{
\item \code{mgcv::cox.ph()} for \link[survival]{Surv} endpoint;
\item \code{binomial(link = 'logit')} for \link[base]{logical} endpoint;
\item \code{gaussian(link = 'identity')} for \link[base]{double} endpoint.
}}

\item{nonlinear}{\link[base]{logical} scalar,
default \code{FALSE} indicating a linear functional regression model is used}

\item{knot_pct}{(only when \code{nonlinear = FALSE})
positive \link[base]{double} scalar,
percentage of the number of columns of \eqn{X},
to be used as \code{knot_value}.
Default is \eqn{40\%}.
If \code{knot_value} is provided by the end-user, then \code{knot_pct} is ignored.}

\item{knot_value}{(only when \code{nonlinear = FALSE})
positive \link[base]{integer} scalar, number of knots
(i.e., parameter \code{k} in the spline smooth function \link[mgcv]{s})
used in \link[mgcv]{gam}.
Default is the \link[base]{ceiling} of \code{knot_pct} of
the column dimension of \eqn{X}}
}
\value{
Function \link{FRidx} returns an \link[base]{S4} class \linkS4class{FRidx} object,
the slots of which are described in section \strong{Slots}.
}
\description{
Functional regression indices
based on linear and/or nonlinear functional predictors.
}
\details{
Function \link{FRidx} calculates
the functional regression indices in the following steps.

\enumerate{

\item Fit a functional regression model (via function \link[mgcv]{gam})
to the response \eqn{y}
using the functional predictor \eqn{X}.

\item Select one point in the tabulating \code{xgrid}.
For one-dimensional domain,
we select the nearest-even \link[stats]{quantile} of the tabulating \code{xgrid},
corresponding to the user-specified probability \code{sign_prob}.
Default \code{sign_prob = .5} indicates the \link[stats]{median} of \code{xgrid}.

\item Obtain the \link[base]{sign} of the \link[stats]{cor}relation between
\itemize{
\item the subject-specific functional predictor values \eqn{X},
at the selected quantile of \code{xgrid} (from Step 2), and
\item the subject-specific \code{gam(.)$linear.predictors}
}

}

\emph{Functional regression indices} (slot \verb{@.Data})
are the product of
\code{sign} (from Step 3) and \code{gam(.)$linear.predictors}.
Multiplication by \code{sign} is required to ensure
that the resulting functional regression indices
are positively associated with the functional predictor values
at the selected quantile of \code{xgrid} (from Step 2).
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{double} \link[base]{vector},
functional regression indices, see section \strong{Details}.}

\item{\code{formula}}{see explanations in section \strong{Arguments}}

\item{\code{gam}}{a \link[mgcv]{gam} object}

\item{\code{gpf}}{a \code{'gam.prefit'} object, which is the returned object
from function \link[mgcv]{gam} with argument \code{fit = FALSE}}

\item{\code{p.value}}{\link[base]{numeric} scalar,
\eqn{p}-value for the test of significance of the functional predictor,
based on slot \verb{@gam}}

\item{\code{sign}}{\link[base]{double} scalar of either 1 or -1,
see Step 3 in section \strong{Details}.}
}}

\examples{
# see ?`Qindex-package`
}
\references{
Cui, E., Crainiceanu, C. M., & Leroux, A. (2021).
Additive Functional Cox Model. Journal of Computational and Graphical Statistics.
\doi{10.1080/10618600.2020.1853550}

Gellar, J. E., Colantuoni, E., Needham, D. M., & Crainiceanu, C. M. (2015).
Cox regression models with functional covariates for survival data. Statistical Modelling.
\doi{10.1177/1471082X14565526}
}
