% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optQp.R
\name{optQp}
\alias{optQp}
\title{\link{optQp}}
\usage{
optQp(x, n = 3L, ...)
}
\arguments{
\item{x}{\link{evalQp} object}

\item{n}{\link[base]{integer} scalar}

\item{...}{additional parameters, not currently in use}
}
\value{
\link{optQp} returns ..
}
\description{
\link{optQp}
}
\examples{
# see ?BBC_dichotom

}
