% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_info.R
\docType{package}
\name{Qindex-package}
\alias{Qindex-package}
\title{Continuous and Dichotomized Index Predictors Based on Distribution Quantiles}
\description{
Primary functions in this package are

\describe{

\item{\code{\link[=optimSplit_dichotom]{optimSplit_dichotom()}}}{
optimal dichotomizing predictor(s) selection via dichotomizing split sample
\strong{Still need? Select optimal functional regression or dichotomized quantile predictors for survival/logistic/numeric outcome}
}

\item{\code{\link[=BBC_dichotom]{BBC_dichotom()}}}{
Bootstrap-based optimism correction for dichotomizing selected predictor(s)
\strong{Still need? perform optimism correction for any optimal dichotomizing predictor(s)}
}

\item{\code{\link[=clusterQp]{clusterQp()}}}{calculate user-selected sample quantiles in each cluster of observations.}

\item{\code{\link[=FRindex]{FRindex()}}}{Functional regression index as a predictor in the functional regression model}

}
}
\references{
Selection of optimal quantile protein biomarkers based on cell-level immunohistochemistry data.
Misung Yi, Tingting Zhan , Amy P. Peck, Jeffrey A. Hooke, Albert J. Kovatich, Craig D. Shriver,
Hai Hu, Yunguang Sun, Hallgeir Rui and Inna Chervoneva.  Under revision

Quantile index biomarkers based on single-cell expression data.
Misung Yi, Tingting Zhan, Amy P. Peck, Jeffrey A. Hooke, Albert J. Kovatich, Craig D. Shriver,
Hai Hu, Yunguang Sun, Hallgeir Rui and Inna Chervoneva.
Laboratory Investigation, 2023. \doi{10.1016/j.labinv.2023.100158}
}
\keyword{package}
