% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0info.R
\docType{package}
\name{Qindex-package}
\alias{Qindex-package}
\title{Qindex}
\description{
Optimize quantile-based predictors for survival outcome (and includes the means to handle dichotomizing the MSI or other continuous markers).
}
\details{
Will provide later
}
\references{
paper in review
}
\keyword{package}
